% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{add_task}
\alias{add_task}
\title{Add an item to a todo list}
\usage{
add_task(todo_name, base_url, task)
}
\arguments{
\item{todo_name}{Name of the todo list}

\item{base_url}{Base url of the todo api}

\item{task}{Task to add}
}
\description{
Add an item to a todo list
}
