% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{delete_task}
\alias{delete_task}
\title{Delete an item from a todo list}
\usage{
delete_task(todo_name, base_url, task_id)
}
\arguments{
\item{todo_name}{Name of the todo list}

\item{base_url}{Base url of the todo api}

\item{task_id}{Id of the task to delete}
}
\description{
Delete an item from a todo list
}
