% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{get_todo_list}
\alias{get_todo_list}
\title{Get all items in a todo list}
\usage{
get_todo_list(todo_name, base_url)
}
\arguments{
\item{todo_name}{Name of the todo list}

\item{base_url}{Base url of the todo api}
}
\value{
A named list
}
\description{
Get all items in a todo list
}
