% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{update_task}
\alias{update_task}
\title{Update a todo list item}
\usage{
update_task(todo_name, base_url, task_id, updated_item)
}
\arguments{
\item{todo_name}{Name of the todo list}

\item{base_url}{Base url of the todo api}

\item{task_id}{Id of the task to update}

\item{updated_item}{The new updated item}
}
\description{
Update a todo list item
}
